Refactored Navbar and Footer into their own components outside of the maing landing page.
